feat(system): 添加系统基础实体和数据访问层
- 新增AppUser、SysConfig、SysDictData等核心实体类
- 添加对应的Mapper接口包括AppUserMapper、SysConfigMapper等
- 配置MybatisPlus乐观锁拦截器
- 创建代码生成器工具类
- 添加系统用户、角色、菜单等基础数据实体
- 生成操作日志、访问记录等实体及Mapper
- 实现行政区划、字典类型等系统配置实体
- 完善用户角色关联、角色菜单关联等多对多关系实体

+++ app/src/test/java/com/xiyao/app/AdminApplicationTests.java
@@ -1,7 +1,6 @@
 package com.xiyao.app;
 
 import com.baomidou.mybatisplus.core.conditions.update.LambdaUpdateWrapper;
-import com.baomidou.mybatisplus.extension.toolkit.Db;
 import com.xiyao.system.entity.AppUser;
 import com.xiyao.system.mapper.AppUserMapper;
 import org.junit.jupiter.api.Test;
@@ -23,12 +22,12 @@ class AdminApplicationTests {
         user.setOpenid("wx_openid_123456");
         user.setName("张三");
         user.setPhone("13812345678");
-        user.setSex((byte) 1);  // 1-男
+        user.setSex(1);  // 1-男
         user.setAvatar("https://example.com/avatar.jpg");
         user.setRegions("广东省深圳市");
         user.setLoginIp("192.168.1.100");
         user.setLoginDate(LocalDateTime.now());
-        user.setStatus((byte) 1);  // 1-正常
+        user.setStatus(1);  // 1-正常
         user.setRemark("测试用户");
 
         int rows = appUserMapper.insert(user);
